factor common code into util helper file
